backend/requirements.txt
flask==2.3.0
flask-cors==3.0.10
flask-jwt-extended==4.5.3
pymongo==4.6.1
python-dotenv==1.0.0
gtts==2.4.0
transformers==4.40.1
torch==2.2.1
sentencepiece==0.2.0
protobuf==4.25.3
youtube-transcript-api==0.6.1
numpy==1.26.4
PyPDF2==3.0.1
newspaper3k==0.2.8
lxml==5.2.1
lxml_html_clean==0.1.1
deep-translator==1.11.4
requests==2.31.0
reportlab==4.1.0
werkzeug==3.0.1
gunicorn
google-generativeai==0.4.0